+++ packages/loader/src/MaterialLoader.ts
@@ -10,12 +10,12 @@ import {
   resourceLoader
 } from "@galacean/engine-core";
 import { Color, Vector2, Vector3, Vector4 } from "@galacean/engine-math";
-import { IPrefabMaterial } from "./resource-deserialize/resources/prefab/PrefabDesign";
+import { IAssetRef, IColor, IMaterialSchema, IVector2, IVector3, IVector4 } from "./resource-deserialize";
 
-function setProperty(object: Object, key: string, value: any) {
+function parseProperty(object: Object, key: string, value: any) {
   if (typeof value === "object") {
     for (let subKey in value) {
-      setProperty(object[key], subKey, value[subKey]);
+      parseProperty(object[key], subKey, value[subKey]);
     }
   } else {
     object[key] = value;
@@ -30,37 +30,46 @@ class MaterialLoader extends Loader<Material> {
         ...item,
         type: "json"
       })
-        .then((json: IPrefabMaterial) => {
+        .then((materialSchema: IMaterialSchema) => {
           const engine = resourceManager.engine;
-          const { name, shader, shaderData, macros, renderState } = json;
+          const { name, shader, shaderData, macros, renderState } = materialSchema;
           const material = new Material(engine, Shader.find(shader));
           material.name = name;
 
-          const texturePromises = new Array<Promise<Texture2D | void>>();
+          const texturePromises = new Array<Promise<Texture2D>>();
           const materialShaderData = material.shaderData;
           for (let key in shaderData) {
             const { type, value } = shaderData[key];
 
             switch (type) {
               case "Vector2":
-                materialShaderData.setVector2(key, new Vector2(value.x, value.y));
+                materialShaderData.setVector2(key, new Vector2((<IVector2>value).x, (<IVector2>value).y));
                 break;
               case "Vector3":
-                materialShaderData.setVector3(key, new Vector3(value.x, value.y, value.z));
+                materialShaderData.setVector3(
+                  key,
+                  new Vector3((<IVector3>value).x, (<IVector3>value).y, (<IVector3>value).z)
+                );
                 break;
               case "Vector4":
-                materialShaderData.setVector4(key, new Vector4(value.x, value.y, value.z, value.w));
+                materialShaderData.setVector4(
+                  key,
+                  new Vector4((<IVector4>value).x, (<IVector4>value).y, (<IVector4>value).z, (<IVector4>value).w)
+                );
                 break;
               case "Color":
-                materialShaderData.setColor(key, new Color(value.r, value.g, value.b, value.a));
+                materialShaderData.setColor(
+                  key,
+                  new Color((<IColor>value).r, (<IColor>value).g, (<IColor>value).b, (<IColor>value).a)
+                );
                 break;
               case "Float":
-                materialShaderData.setFloat(key, value);
+                materialShaderData.setFloat(key, <number>value);
                 break;
               case "Texture":
                 texturePromises.push(
                   // @ts-ignore
-                  resourceManager.getResourceByRef<Texture2D>(value).then((texture) => {
+                  resourceManager.getResourceByRef<Texture2D>(<IAssetRef>value).then((texture) => {
                     materialShaderData.setTexture(key, texture);
                   })
                 );
@@ -77,7 +86,7 @@ class MaterialLoader extends Loader<Material> {
             }
           }
 
-          setProperty(material, "renderState", renderState);
+          parseProperty(material, "renderState", renderState);
 
           return Promise.all(texturePromises).then(() => {
             resolve(material);
