+++ packages/loader/src/resource-deserialize/resources/scene/SceneParser.ts
@@ -1,8 +1,8 @@
 import { Engine, Entity, Loader, Scene } from "@galacean/engine-core";
-import { IEntity, IScene } from "../prefab/PrefabDesign";
-import { PrefabParser } from "../prefab/PrefabParser";
-import { ReflectionParser } from "../prefab/ReflectionParser";
 import { SceneParserContext } from "./SceneParserContext";
+import { IScene } from "../schema";
+import { ReflectionParser } from "../parser/ReflectionParser";
+import { PrefabParser } from "../parser/PrefabParser";
 
 /** @Internal */
 export class SceneParser {
