+++ packages/loader/src/resource-deserialize/resources/schema/index.ts
@@ -1,0 +1,3 @@
+export * from "./BasicSchema";
+export * from "./MaterialSchema";
+export * from "./SceneSchema";
